add nd4j cast and a small test for yolo
